| GroupId | ReleaseVersion | ReleaseSHA | ReleaseBuildId | ReleaseDate |
| --- | --- | --- | --- | --- |
| androidx.compose.animation | 1.11.0-alpha05 |
cdf076c6abd0f3125cb0302756fcb77fe981ab7c | 14838085 | 2/11/2026 |
| androidx.compose.foundation | 1.11.0-alpha05 |
cdf076c6abd0f3125cb0302756fcb77fe981ab7c | 14838085 | 2/11/2026 |
| androidx.compose.material |1.11.0-alpha05 |
cdf076c6abd0f3125cb0302756fcb77fe981ab7c | 14838085 | 2/11/2026 |
| androidx.compose.runtime | 1.11.0-alpha05 |
cdf076c6abd0f3125cb0302756fcb77fe981ab7c | 14838085 | 2/11/2026 |
| androidx.compose.ui | 1.11.0-alpha05 |
cdf076c6abd0f3125cb0302756fcb77fe981ab7c | 14838085 | 2/11/2026 |

Fixes
https://youtrack.jetbrains.com/issue/CMP-9742/Merge-Jetpack-Compose-1.11.0-alpha05

## Testing
N/A

## Release Notes
N/A

+++ compose/remote/remote-creation-compose/src/main/java/androidx/compose/remote/creation/compose/modifier/VisibilityModifier.kt
@@ -25,7 +25,7 @@ import androidx.compose.remote.creation.modifiers.RecordingModifier
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 public class VisibilityModifier(public val visible: RemoteInt) : RemoteModifier.Element {
     override fun RemoteStateScope.toRecordingModifierElement(): RecordingModifier.Element {
-        return androidx.compose.remote.creation.modifiers.VisibilityModifier(visible.id.toInt())
+        return androidx.compose.remote.creation.modifiers.VisibilityModifier(visible.id)
     }
 }
